+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/custom.py
@@ -2618,12 +2618,13 @@ def update_image_version(instance, target_regions=None, replica_count=None):
 
 # region Proximity Placement Group
 def create_proximity_placement_group(cmd, client, proximity_placement_group_name, resource_group_name,
-                                     ppg_type, location=None):
+                                     ppg_type, location=None, tags=None):
     location = location or _get_resource_group_location(cmd.cli_ctx, resource_group_name)
 
     ProximityPlacementGroup = cmd.get_models('ProximityPlacementGroup')
 
-    ppg_params = ProximityPlacementGroup(name=proximity_placement_group_name, type=ppg_type, location=location)
+    ppg_params = ProximityPlacementGroup(name=proximity_placement_group_name, proximity_placement_group_type=ppg_type,
+                                         location=location, tags=(tags or {}))
 
     return client.create_or_update(resource_group_name=resource_group_name,
                                    proximity_placement_group_name=proximity_placement_group_name, parameters=ppg_params)